backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
anthropic==0.7.7
openai==1.3.5
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
requests==2.31.0
beautifulsoup4==4.12.2
fpdf2==2.7.6
reportlab==4.0.4
openpyxl==3.1.2
pandas==2.1.3
numpy==1.25.2
Pillow==10.1.0
python-dotenv==1.0.0
dj-database-url==2.1.0

